OpenCV_Hough_Circle_Transformation_Coins
This program uses the Hough Circle Transform in OpenCV to detect coins  (one coin with a different shape is omitted).  Their radii is then used to calculate the amount of money. The output picture depicts outlines of each coin with labels and the estimated value of all coins. The code is a modified version of the code available in the OpenCV tutorial on Hough Circle Transformation.